requirements.txt
streamlit==1.32.0
langchain==0.1.16
langchain-groq==0.1.3
langchain-community==0.0.36
chromadb==0.4.24
sentence-transformers==2.7.0
pandas==2.2.1
python-dotenv==1.0.1
datasets==2.18.0
